Move ddl compilation to separate test file
